# protocol/openapi.yaml
openapi: 3.0.3
info:
  title: Remote File Explorer — Host Agent API
  version: 1.17.0
  description: >
    Contract between the Flutter mobile app and the Go host agent. JSON for control and
    metadata; binary streams for file content. All routes except /pair, /register,
    /login, /auth/challenge, and /health require a bearer device token. Clients may
    send an optional `X-RFE-Client-Version` header (e.g.
    "1.10.0+18") on every authenticated request; the agent records this plus the
    caller's network address against the device row. Admin devices see those fields
    for all devices via GET /devices (see Device.lastAddress / Device.lastVersion);
    ordinary paired devices receive only their own minimal device record.

    A device's token is "admin" (Device.viaLogin=true) if it was obtained via /login or
    /register (proof of the single account's password) rather than /pair (a one-time
    code meant for an ordinary device like the phone app). Admin devices may manage
    OTHER devices: POST /pairing/generate, and PATCH/DELETE /devices/{id} targeting a
    device other than themselves. An ordinary (non-admin) device may still manage only
    ITSELF via DELETE /devices/{id}.

    Pair, register, and login all require device-identity proof-of-possession: fetch a
    nonce from POST /auth/challenge, sign it with the device's permanent Ed25519 key,
    and send the (devicePublicKey, nonce, signature) triple. The agent pins the public
...
servers:
...
    variables:
...
security:
  - deviceToken: []
paths:
...
  /transfers:
    post:
      summary: Open a resumable upload session
      description: >
        Declared upload size is capped at 64 GiB. Admission reserves the full
        declared size: at most 4 open sessions and 64 GiB per device, and 16
        open sessions and 128 GiB across the host. Reaching either session or
        byte quota returns 429 RESOURCE_LIMIT. Reservations bound RFE's
        concurrent declared uploads; they do not guarantee free disk space or
        account for unrelated host writes. Open sessions with no chunk or
        status/resume activity expire after 7 days; the agent sweeps hourly.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path, size, sha256, chunkSize]
              properties:
                path: { type: string }
                size:
                  type: integer
                  format: int64
                  minimum: 0
                  maximum: 68719476736
                  description: Maximum 64 GiB per upload.
                sha256: { type: string }
                chunkSize:
                  type: integer
                  minimum: 1
                  maximum: 33554432
                  description: Maximum 32 MiB per chunk.
                overwrite:
                  type: boolean
                  default: false
                  description: >
                    When false (default) and path already exists, the
                    session is rejected with 409 CONFLICT. The flag is
                    recorded on the session and re-checked atomically at
                    POST /transfers/{id}/complete, so a file that appears
                    during the upload also conflicts (409) instead of being
                    silently replaced. When true, an existing file at path is
                    replaced once the upload completes.
      responses:
        "201":
          description: Session opened
          content: { application/json: { schema: { $ref: "#/components/schemas/UploadSession" } } }
        "400":
          description: >
            Invalid request — missing path/sha256/chunkSize, negative size, or
            chunkSize over the 32MiB cap. code=BAD_REQUEST.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "401": { $ref: "#/components/responses/Error" }
        "403":
          description: >
            Path outside the caller's jail, or the agent/device is read-only.
            code=FORBIDDEN or READ_ONLY. Every mutating transfer route honours
            read-only, the same as the /fs routes.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "409":
          description: >
            path already exists and overwrite=false. code=CONFLICT.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "413":
          description: Declared upload exceeds the 64 GiB per-upload maximum. code=PAYLOAD_TOO_LARGE.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "429":
          description: >
            Per-device or host-wide active-session/reserved-byte quota is full.
            Finish or delete an existing session before opening another.
            No Retry-After is supplied because capacity can become available
            as soon as another upload finishes. code=RESOURCE_LIMIT.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /transfers/{id}:
    get:
      summary: Get session status + missing chunks (for resume)
      description: >
        A successful status/resume request by the session owner refreshes the
        session's activity time. Open sessions with no chunk or owner
        status/resume activity for 7 days are eligible for the agent's hourly
        stale-session sweep.
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses:
        "200": { description: Session, content: { application/json: { schema: { $ref: "#/components/schemas/UploadSession" } } } }
        "401": { $ref: "#/components/responses/Error" }
        "409":
          description: Completion currently holds the session lease. code=TRANSFER_ACTIVE.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      summary: Delete an inactive transfer session (web companion Transfers page)
      description: >
        Deletes the session record and its temporary upload file; it never
        deletes a file already published at the destination. Deletion is
        coordinated with chunk uploads, completion, and stale cleanup. If a
        session operation is active, deletion returns 409 TRANSFER_ACTIVE and
        removes nothing. An old unreferenced `.tmp` file left after a previous
        deletion failure is removed by age-gated cleanup once it is older than
        the 7-day retention period.

        Scoped to the session's owner: only the device that opened it, or an
        admin device, may remove the row. Anyone else gets 404 — the same
        response an unknown ID gets, so a foreign session ID is never
        confirmed. Rows with no recorded owner are admin-only.
      security: [{ deviceToken: [] }]
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses:
        '204': { description: Transfer row removed }
        '401': { $ref: '#/components/responses/Error' }
        '409':
          description: An active session operation currently holds the transfer. code=TRANSFER_ACTIVE.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        '404':
          description: >
            No transfer with that ID, or it belongs to another device.
            code=NOT_FOUND.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
components:
  securitySchemes:
    deviceToken:
      type: http
      scheme: bearer
  responses:
    Error:
      description: Error
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
...
    UploadSession:
      type: object
      properties:
        id: { type: string }
        path: { type: string }
        size: { type: integer, format: int64 }
        chunkSize: { type: integer }
        totalChunks: { type: integer }
        receivedChunks: { type: array, items: { type: integer }, description: Indices already stored }
        status: { type: string, enum: [open, completed, failed] }
...
    Error:
      type: object
      description: >
        Uniform error envelope returned on non-2xx responses. The `code` field is
        the authoritative machine-readable vocabulary the agent emits (see enum);
        `message` is a human-readable detail string and may change without notice.
      properties:
        code:
          type: string
          example: PATH_NOT_FOUND
          enum:
            - BAD_REQUEST
            - UNAUTHORIZED
            - READ_ONLY
            - FORBIDDEN
            - PATH_NOT_FOUND
            - CONFLICT
            - CHUNK_HASH_MISMATCH
            - HASH_MISMATCH
            - NOT_AVAILABLE
            - THUMB_BUSY
            - INVALID_CODE
            - DELETE_FAILED
            - RATE_LIMITED
            - RESOURCE_LIMIT
            - TRANSFER_ACTIVE
            - TRANSFER_NOT_OPEN
            - PAYLOAD_TOO_LARGE
            - NOT_FOUND
            - INTERNAL
          description: >
            - `BAD_REQUEST` (400): malformed or missing required fields.
            - `UNAUTHORIZED` (401): missing/invalid/revoked bearer token.
            - `READ_ONLY` (403): write rejected — the agent is in global
              read-only mode OR the calling device has the per-device read-only
              flag set (#8).
            - `FORBIDDEN` (403): path escapes the configured root jail; or a
              /devices/{id} DELETE/PATCH or /pairing/generate call that
              requires an admin device (Device.viaLogin=true) came from a
              non-admin device.
            - `PATH_NOT_FOUND` (404): path does not exist.
            - `NOT_FOUND` (404): PATCH /devices/{id} — no device with that id.
            - `CONFLICT` (409): destination already exists.
            - `CHUNK_HASH_MISMATCH` (409): per-chunk SHA-256 did not match
              X-Chunk-Sha256.
            - `HASH_MISMATCH` (422): whole-file SHA-256 did not match on complete.
            - `NOT_AVAILABLE` (404): no thumbnail/preview available for this file.
            - `THUMB_BUSY` (429): thumbnail request capacity is full; retry after the `Retry-After` interval.
            - `INVALID_CODE` (401): pairing code invalid or expired.
            - `DELETE_FAILED`: per-item batch-delete failure (returned inside a
              BatchResult entry, not as a top-level HTTP error).
            - `RATE_LIMITED` (429): too many /pair attempts (10/min).
            - `RESOURCE_LIMIT` (429): upload admission reached the per-device
              (4 sessions / 64 GiB) or host-wide (16 sessions / 128 GiB)
              session or reserved-byte quota. Retry after completing or
              deleting an existing session; no fixed Retry-After is provided.
            - `TRANSFER_ACTIVE` (409): an operation conflicts with active
              session work; completion is exclusive, and DELETE is rejected
              while another request holds the session.
            - `TRANSFER_NOT_OPEN` (409): the session is completed or failed
              and no longer accepts chunks or completion.
            - `PAYLOAD_TOO_LARGE` (413): chunk body exceeds the session's chunkSize.
            - `INTERNAL` (500): unexpected server error.
        message: { type: string }
        details: { type: object, additionalProperties: true }